refactor(network): refactor the network module into more manageable smaller files
All functionality and comments are preserved
