+++ rs/decentralization/src/network/mod.rs
@@ -1,0 +1,25 @@
+// Re-export all public items from submodules
+mod change;
+mod healing;
+mod request;
+mod subnet;
+mod traits;
+mod types;
+
+pub use change::*;
+pub use healing::*;
+pub use request::*;
+pub use subnet::*;
+pub use traits::*;
+pub use types::*;
+
+// Common imports used across the network module
+pub(crate) use ahash::{AHashMap, AHashSet, HashSet};
+pub(crate) use anyhow::anyhow;
+pub(crate) use futures::future::BoxFuture;
+pub(crate) use ic_base_types::PrincipalId;
+pub(crate) use ic_management_types::{HealthStatus, NetworkError, Node, NodeFeature};
+pub(crate) use indexmap::{IndexMap, IndexSet};
+pub(crate) use itertools::Itertools;
+pub(crate) use rand::{seq::SliceRandom, SeedableRng};
+pub(crate) use serde::{Deserialize, Serialize};
